+++ package.json
@@ -2,7 +2,7 @@
   "name": "fast-graphql-mock",
   "version": "0.0.1",
   "description": "A library for mocking GraphQL schema using one command",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
